+++ src/main/java/io/suhan/lotto/model/executor/LottoPurchaseExecutor.java
@@ -1,10 +1,10 @@
 package io.suhan.lotto.model.executor;
 
-import io.suhan.lotto.model.LottoFactory;
-import io.suhan.lotto.model.LottoRegistry;
+import io.suhan.lotto.model.lotto.LottoFactory;
+import io.suhan.lotto.model.lotto.LottoRegistry;
 
 public class LottoPurchaseExecutor implements LottoExecutor {
-    private static final int PRICE_PER_LOTTO = 1000;
+    public static final int PRICE_PER_LOTTO = 1000;
     private final LottoRegistry registry;
     private final int balance;
