+++ src/main/java/io/suhan/lotto/model/lotto/Lotto.java
@@ -1,4 +1,4 @@
-package io.suhan.lotto.model;
+package io.suhan.lotto.model.lotto;
 
 import java.util.Comparator;
 import java.util.HashSet;
@@ -6,6 +6,9 @@ import java.util.Set;
 
 public class Lotto {
     public static final int LOTTO_SIZE = 6;
+    public static final int LOTTO_NUMBER_MIN = 1;
+    public static final int LOTTO_NUMBER_MAX = 45;
+
     private final Set<LottoNumber> numbers;
 
     public Lotto(Set<LottoNumber> numbers) {
